+++ lib/glimmer/libui.rb
@@ -19,15 +19,23 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+require 'glimmer/fiddle_consumer'
+
 module Glimmer
   module LibUI
     class << self
+      include Glimmer::FiddleConsumer
+      
       def integer_to_boolean(int, allow_nil: true)
-        int.nil? ? (allow_nil ? nil : false) : int == 1
+        int.nil? ? (allow_nil ? nil : false) : ((int.is_a?(TrueClass) || int.is_a?(FalseClass)) ? int : int == 1)
       end
       
       def boolean_to_integer(bool, allow_nil: true)
-        bool.nil? ? (allow_nil ? nil : 0) : (bool ? 1 : 0)
+        bool.nil? ? (allow_nil ? nil : 0) : (bool.is_a?(Integer) ? bool : (bool == true ? 1 : 0))
+      end
+      
+      def degrees_to_radians(degrees)
+        ((Math::PI * 2.0) / 360.00) * degrees.to_f
       end
       
       def interpret_color(value)
@@ -103,6 +111,43 @@ module Glimmer
         else
           enum_symbol_to_value(enum_name, enum_symbols(enum_name)[default_index])
         end
+      end
+      
+      def x11_colors
+        Color::RGB.constants.reject {|c| c.to_s.upcase == c.to_s}.map(&:to_s).map(&:underscore).map(&:to_sym)
+      end
+      
+      # Queues block to execute at the nearest opportunity possible on the main GUI event loop
+      def queue_main(&block)
+        closure = fiddle_closure_block_caller(4, [0]) do
+          result = boolean_to_integer(block.call)
+          result = 1 if result.nil?
+          result
+        end
+        ::LibUI.queue_main(closure)
+        closure
+      end
+      
+      # Calls block on the main GUI event loop after time_in_seconds delay, repeating indefinitely by default
+      # If `repeat:` keyword arg is passed with an Integer value, it repeats for that number of times
+      # If `repeat:` keyword arg is passed with false or 0, then the block is only called onces
+      # If block returns false at any point, the timer is stopped from further repetitions regardless of `repeat:` keyword arg value
+      # If block returns true at any point, the timer continues for another repetition regardless of `repeat:` keyword arg value
+      def timer(time_in_seconds = 0.1, repeat: true, &block)
+        closure = fiddle_closure_block_caller(4, [0]) do
+          result = boolean_to_integer(block.call)
+          repeat -= 1 if repeat.is_a?(Integer)
+          if result.nil?
+            if (repeat == true || (repeat.is_a?(Integer) && repeat > 0))
+              result = 1
+            else
+              result = 0
+            end
+          end
+          result
+        end
+        ::LibUI.timer(time_in_seconds * 1000.0, closure)
+        closure
       end
     end
   end